% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{create_pums_survey}
\alias{create_pums_survey}
\title{Create a survey design object from the Census PUMS dataset}
\usage{
create_pums_survey(data_frame, wgt, rep_wgt)
}
\arguments{
\item{data_frame}{A data frame used in the function to create the table.}

\item{wgt}{Name of the weight column, unquoted; either `WGTP` (household) or `PWGTP` (population).}

\item{rep_wgt}{Name of the replicate weight columns as quoted regular expressions.
Either `"WGTP[0-9]+"` (household) or `"PWGTP[0-9]+"` (population)}
}
\description{
This function uses the `srvy` package to create a survey design from the PUMS dataset
with replicate weights.  Survey designs can be creaed for the household or population datasets.
}
