% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{ff_acs_age}
\alias{ff_acs_age}
\title{Filter for age rows in ACS data and rebin ages if needed}
\usage{
ff_acs_age(df, age_column, recode = F, recode_list = NULL)
}
\arguments{
\item{df}{Dataframe of ACS data}

\item{age_column}{Column name that contains age information}

\item{recode}{Boolean value of whether to recode age bins}

\item{recode_list}{A named list where name is old age bin and value is a string for the new age bin}
}
\value{
Dataframe that only contains the age rows
}
\description{
This function filters ACS data for rows that include ages.  It also has optional
parameters to rebin age groups into custom age bins.
}
\examples{
age_bins <- list(`16 to 26 year old` = "16 to 25",
                 `75 and over` = "over 75")
ff_acs_age(df, description, recode = T, recode_list = age_bins)

}
