% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{ff_acs_ethnicity}
\alias{ff_acs_ethnicity}
\title{Filter for race / ethnicity rows in ACS data}
\usage{
ff_acs_ethnicity(df, ethnicity_column)
}
\arguments{
\item{df}{Dataframe of ACS data}

\item{ethnicity_column}{Column name that contains race / ethnicity phrase}
}
\value{
The dataframe inputted into the function, but only the rows with race / ethnicity information
        for White Alone, Not Hispanic or Latino; Black or African American Alone; and Hispanic or Latino remain
}
\description{
This function filters for the following ethnicities in ACS data:
     White Alone, Not Hispanic or Latino; Black or African American Alone; Hispanic or Latino
}
\details{
It also cleans the race / ethnicity name to match FF standards.
}
\examples{
ff_acs_ethnicity(df, description)

}
