% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{ff_acs_keep_vars}
\alias{ff_acs_keep_vars}
\title{Filter ACS data for specific variables}
\usage{
ff_acs_keep_vars(df, variables)
}
\arguments{
\item{df}{Dataframe of ACS data}

\item{variables}{Vector of strings signifying variable numbers}
}
\value{
Same dataframe as input, but only the rows with the needed variables are included.
}
\description{
ACS variables are three digit numbers shown as the last three digits in the 'variables' column.
The function filters a dataframe of ACS data for specified variables.
}
\examples{
ff_acs_keep_vars(df, c('100', '101'))

}
