% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{ff_county_fips}
\alias{ff_county_fips}
\title{Match county names to county FIPS codes}
\usage{
ff_county_fips(county_fips_codes)
}
\arguments{
\item{county_fips}{NC county FIPS codes.}
}
\value{
The county name as a string corresponding to the county FIPS code.
}
\description{
This function performs a cross-walk between county names and county FIPS codes.
It takes NC county FIPS codes and input and outputs the county name.
}
\examples{
df <- data.frame(county_fips_code = c(1, 3, 67),
                 estimate = c(121, 156, 190))

ff_county_fips(df$county_fips_code)

# Using pipes to add county FIPS names as a new column to dataframe
df \%>\%
    mutate(county_name = ff_county_fips(.$county_fips_code))

}
