% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{ff_create_varnames}
\alias{ff_create_varnames}
\title{Create variable names for tables}
\usage{
ff_create_varnames(data_frame, table_data, var_names)
}
\arguments{
\item{data_frame}{A data frame used in the function to create the table.}

\item{table_data}{The significance test or estimate table created by `ff_sigtest` or `ff_estimate_ci`.}

\item{var_names}{The variable names used inthe function to create the table.}
}
\description{
This function creates variable names for tables when `var_names = TRUE` in
`ff_sigtest` and `ff_estimate_ci`. Users will not call this function, it will
be called within `ff_sigtest` and `ff_estimate_ci`.
}
