% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_census.R
\name{ff_import_tables}
\alias{ff_import_tables}
\title{This function returns a dataframe of multiple table numbers at one time}
\usage{
ff_import_tables(geography, table_number, state = NULL, county = NULL,
  year_end, acs_data)
}
\description{
This function returns a dataframe of multiple table numbers at one time
}
