% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_census.R
\name{ff_import_tidycensus}
\alias{ff_import_tidycensus}
\title{Return ACS tables for multiple geographies and years}
\usage{
ff_import_tidycensus(geography, table_number, state = NULL,
  county = NULL, year_start, year_end, acs_data = "acs_acs1")
}
\arguments{
\item{geography}{Geographic level of interest, using notation in tidycensus packages}

\item{table_number}{ACS table number}

\item{state}{Vector of states (must align with vector of counties)}

\item{county}{Vector of counties that align with the vector of states}

\item{year_start}{Start year as integer}

\item{year_end}{End year as integer}

\item{acs_data}{Table type, this is the 'survey' parameter in 'get_acs'.
Use 'acs1' for data from 2015 and later; and use 'acs/acs1' for data prior to 2015}
}
\value{
A dataframe of ACS data
}
\description{
Return ACS tables for multiple geographies and years
}
\examples{
census_data(geography = 'county',
            table_number = 'S1903',
            state = c('NC', 'NC'),
            county = c('Forsyth', 'Guilford'),
            year_start = 2011,
            year_end = 2018)

}
