% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_census.R
\name{ff_iterate_acs}
\alias{ff_iterate_acs}
\title{Iteratively import ACS data}
\usage{
ff_iterate_acs(parameters_list, variable_names = NULL)
}
\arguments{
\item{parameters_list}{A named list of parameters to feed into the get_acs function from
tidycensus.  The name of each item should correspond to the parameter name in get_acs.}

\item{variable_names}{A dataframe that contains the variable names.  Default is NULL.}
}
\value{
A tidy dataframe with all census data.  Additional columns for the yar and geographic
region are added.
}
\description{
This function makes repeated calls to the Census API through the tidycensus package.
It allows users to import multiple years and geographic areas with one function call.
}
\examples{
parameters <- list(geography = c('us', 'state', 'county', 'tract'),
                   state = c(NA, rep('NC', 3)),
                   county = c(rep(NA, 3), 'Forsyth'),
                   table = c(rep('S2301', 3), NA),
                   variables = c(rep(NA, 3), 'S2301_C03_001'),
                   survey = c(rep('acs1', 3), 'acs5'),
                   year = rep(current_year, 4),
                   use_table = c(rep(T, 3), F))

variable_descriptions <- load_variables(2017, "acs1", cache = TRUE)
employment <- iterate_acs(parameters, variable_descriptions)

}
