% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{ff_p_plot}
\alias{ff_p_plot}
\title{Plot of p-values}
\usage{
ff_p_plot(pvalue_matrix)
}
\arguments{
\item{pvalue_matrix}{A table of p-values generated from `ff_sigtest`.}
}
\value{
A ggplot graphic that is color coded based on teh p-value.
}
\description{
This function returns a plot similair to a correlation plot, highlighting whether p-values are significant.
THe function takes as input a table of p-values created by `ff_sigtest`. In using `ff_sigtest` to create a plot
with this function, the `pretty_print` parameter msut be set to false.
}
\examples{
df <- data.frame(year = c(2016, 2016, 2017, 2017),
                 geo_description = c('Forsyth County, NC', 'Guilford County, NC',
                                     'Forsyth County, NC', 'Guilford County, NC'),
                estimate = c(.66, .63, .88, .48),
                se = c(.1, .15, .06, .09),
                success = c(10, 12, 15, 19),
                trials = c(15, 19, 17, 39))

# create table of p-values
pvalues <- ff_sigtest(data_frame = df, estimate = 'estimate', se = 'se',
                      test = 'zscore', var_names = c('year', 'geo_description'))

# generate plot
ff_p_plot(pvalues)

}
