% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{ff_plot_census}
\alias{ff_plot_census}
\title{Choropleth Map for Census Tracts}
\usage{
ff_plot_census(
  data,
  col_census_name,
  col_estimate,
  title,
  reverse_colors = F,
  legend_suffix = NULL,
  legend_prefix = NULL
)
}
\arguments{
\item{data}{Name of the dataframe with data and shapefiles}

\item{col_census_name}{column name as a string of the column with the census tract names}

\item{col_estimate}{column name as a string of the column with the estimate we want to plot}

\item{title}{title of plot}

\item{reverse}{reverse the mapping of the colors to numbers (by default dark colors represent high numbers)}
}
\value{
choropleth map of census tracts and estimate
}
\description{
This function creates choropleth maps for census tract estimates.
}
\examples{
df <- tidycensus::get_acs(state = "NC",
                          county = "Forsyth",
                          geography = "tract",
                          variables = "S2301_C03_001",
                          geometry = TRUE) \%>\%
      dplyr::mutate(NAME = stringr::str_replace(", Forsyth County, North Carolina", ""))

ff_plot_census(df, "NAME", "estimate", "Employment Rate", T, NULL, "\%")

}
