% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{ff_plot_compare_line}
\alias{ff_plot_compare_line}
\title{Line Chart of Comparison Geographies}
\usage{
ff_plot_compare_line(
  data,
  comparison_filter,
  y_axis_title,
  percent = F,
  dollar = F,
  geography_order = c("Forsyth County, NC", "Guilford County, NC", "Durham County, NC",
    "North Carolina", "United States")
)
}
\arguments{
\item{data}{Name of the dataframe with columns outlined above}

\item{comparison_filter}{String in the `type` column that signifies the rows that are needed for this plot,
Will generally be rows of total aggregate value}

\item{y_axis_title}{Title for y axis and tooltip}

\item{percent}{Boolean (TRUE / FALSE), whether estimate value is a percent; this will add percent labels to axis}

\item{dollar}{Boolean (TRUE / FALSE), whether estimate value is a dollar; this will add dollar labels to the axis}

\item{geography_order}{Vector of strings that sets the geography order of the legend}
}
\value{
line chart of all years data
}
\description{
This function creates a line chart of Forsyth County and comparison communities for a given metric.
}
\details{
The input dataset needs to have columns of the following form, and with the following names:
- year: year of the data
- geo_description: geography
- type: general description of the demographic identifier (Race and Ethnicity, Total Population, etc)
- subtype: specific description of the demographic identifier (African American, Employment Rate, etc)
- estimate: value
}
\examples{
geographies <- c("Forsyth County, NC", "Guilford County, NC", "Durham County, NC",
                 "North Carolina", "United States")
years <- seq(2006, 2017, 1)

df <- data.frame(
    year = rep(years, each=length(geographies)),
    geo_description = rep(geographies, length(years)),
    type = "Total Population",
    subtype = "Employment Rate",
    stringsAsFactors = FALSE
)

df$estimate <- rnorm(nrow(df), mean = .5, sd = .15)

ff_plot_compare_line(data = df,
                    comparison_filter = "Total Population",
                    y_axis_title = "Employment Rate (\%)",
                    percent = T,
                    dollar = F,
                    geography_order = c("Forsyth County, NC", "Guilford County, NC", "Durham County, NC",
                                         "North Carolina", "United States"))
}
