% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{ff_plot_demo_bar}
\alias{ff_plot_demo_bar}
\title{Bar Chart of Demographics}
\usage{
ff_plot_demo_bar(
  data,
  demographic,
  y_axis_title,
  percent = F,
  dollar = F,
  color_palette = "Set 2"
)
}
\arguments{
\item{data}{Name of the dataframe with columns outlined above}

\item{demographic}{String in the `type` column that signifies the demographic}

\item{y_axis_title}{Title for y axis and tooltip}

\item{percent}{Boolean (TRUE / FALSE), whether estimate value is a percent; this will add percent labels to axis}

\item{dollar}{Boolean (TRUE / FALSE), whether estimate value is a dollar; this will add dollar labels to the axis}

\item{color_palette}{string representing the color palette to use, possible palettes are qualitative
palettes from the <a href="http://library http://colorspace.r-forge.r-project.org/reference/hcl_palettes.html">`colorspace` library</a>.}
}
\value{
bar plot of the most recent year of data
}
\description{
This function creates a bar chart of a given demographic for the most recent year within Forsyth County.
}
\details{
The input dataset needs to have, at a minimum, columns of the following form and with the following names:
- year: year of the data
- geo_description: geography (needs Forsyth County as one of the geographies)
- type: general description of the demographic identifier (Race and Ethnicity, Total Population, etc)
- subtype: specific description of the demographic identifier (African American, Employment Rate, etc)
- estimate: value
}
\examples{
years <- seq(2006, 2017, 1)

df <- data.frame(
   year = rep(seq(2007, 2017, 1), each=3),
   geo_description = "Forsyth County, NC",
   type = "Race and Ethnicity",
   subtype = rep(c("African American", "White", "Hispanic / Latino"),11),
   stringsAsFactors = FALSE
)

ff_plot_demo_bar(data = df,
                 demographic = "Race and Ethnicity",
                 y_axis_title = "Employment Rate (\%)",
                 percent = T,
                 dollar = F)
}
