% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{ff_pretty_kable}
\alias{ff_pretty_kable}
\title{Pretty formatting of significance testing and estimate tables with Kable}
\usage{
ff_pretty_kable(
  data_matrix,
  table_type,
  format = "continuous",
  table_name = NULL
)
}
\arguments{
\item{data_matrix}{A matrix produced by `ff_sigtest` or `ff_estimates_ci`.}

\item{table_type}{Either 'sigtest' or 'estimate'. Specifies whether the table was generated by
`ff_sigtest` and has results from a significance test; or whether the table was generated
by `ff_estimates_ci` and has estimates and confidence intervals.}

\item{format}{Type of data; either 'continuous' or 'binomial'. Defaults to 'continuous'.}

\item{table_name}{A string of characters representing the table name. It is displayed above the table.
Defautls to no table name.}
}
\value{
An html-based Kable table.
}
\description{
This function elegantly formats significance testing matrices produces
by `ff_sigtest` or estimate matrices produced by `ff_estimates_ci`.
It bolds cells that are statistically significant (for significance testing matrices), only
keeps rows for Forsyth County, and makes other minor stylistic changes.
}
