% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{ff_proportions}
\alias{ff_proportions}
\title{Difference and confidence interval of difference between two binomial observations}
\usage{
ff_proportions(successes, trials)
}
\arguments{
\item{trials}{Vector of two numbers that represent the number of trials in each
of the observations}

\item{success}{Vector of two numbers that represent the number of successes in each
of the observations.}
}
\description{
This function calculates the percentile difference between two observations that
are binomial (true / false, yes / no, etc) in nature. It also calculates the confidence
interval of this difference. The function is used in `ff_estiamtes_ci`.
}
