% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{ff_pums_ethnicity}
\alias{ff_pums_ethnicity}
\title{Relabel PUMS data from numeric race / ethnicity to descriptive string}
\usage{
ff_pums_ethnicity(df)
}
\arguments{
\item{df}{Dataframe of PUMS data.  The dataframe must have the following two columns: `RAC1P` and `HISP`}
}
\value{
The same dataframe, but with the race / ethnicity column `RAC1P` labeled descriptively.
}
\description{
This function labels Race / ethnicities in PUMS data as either White, African American, or
Hispanic / Latino.  The
}
\examples{
ff_pums_ethnicity(df)

}
