% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustments.R
\name{ff_se_count}
\alias{ff_se_count}
\title{Standard error for aggregated count data}
\usage{
ff_se_count(se_vec)
}
\arguments{
\item{se_vec}{A numeric vector of values representing standard errors of count data}
}
\value{
A numeric value representing the standard error of the aggregate count data.
}
\description{
This function calculates the standard error for aggregated count data and summed data. For example,
this can be used to calculate the standard error for the total population of Forsyth and Guilford
counties.
}
\details{
Reference for calculations:
US Census Bureau, A Compass for Understanding and Using ACS Data, October 2008, A-14
}
\examples{
df <- data.frame(population = round( rnorm(n = 10, mean = 100000, sd = 30000), 0),
                 se = round( rnorm(n = 10, mean = 4000, sd = 1500), 0))

# calculate the aggregate population
sum(df$population)

# calculate the aggregate standard error
ff_se_count(df$se)

}
