% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustments.R
\name{ff_se_prod}
\alias{ff_se_prod}
\title{Standard error for the product of two estimates}
\usage{
ff_se_prod(estimate_first, se_first, estimate_second, se_second)
}
\arguments{
\item{estimate_first}{Point estimate of the first value.}

\item{se_first}{Standard error for the first estimate.}

\item{estimate_second}{Point estimate of the second value.}

\item{se_second}{Standard error for the second estimate}
}
\value{
The standard error of the derived products.
}
\description{
This function calculates the standard error for the product of two estimates. An example is if users want
to derive the number of people working by multiplying the total population by the percentage of the
population that is working.
}
\details{
Reference for calculations:
US Census Bureau, A Compass for Understanding and Using ACS Data, October 2008, A-16
}
\examples{
df <- data.frame(pop = rnorm(n = 10, mean = 100000, sd = 20000),
                se_pop = rnorm(n = 10, mean = 1000, sd = 200),
                perc_working = runif(n = 10, min = .6, max = .9),
                perc_working_se = runif(n = 10, min = .05, max = .2))

# calculate standard errors of the products for each observation
ff_se_ratio(df$pop, df$se_pop,
           df$perc_working, df$perc_working_se)

# add columns to dataframe showing products and standard errors of products
dplyr::mutate(df,
              total_working = pop / perc_working,
              se_total_working = ff_se_prod(pop, se_pop,
                                            perc_working, perc_working_se))

}
