% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{ff_sigtest}
\alias{ff_sigtest}
\title{Matrix of significance tests}
\usage{
ff_sigtest(data_frame, estimate, se, test = "z", success = NULL,
  trials = NULL, var_names = NULL)
}
\arguments{
\item{data_frame}{A dataframe containing estimates and either standard errors for z-score test
or successes and trials for a Chi-Square test.}

\item{estimate}{An integer or float containing the number to conduct significance tests on.}

\item{se}{Standard error of the estimate. Required if test a z-score test}

\item{test}{The significance test to conduct. Either "zscore" or "chi-square". Defaults to 'zscore'.}

\item{success}{The number of successful trials. Required for Chi-Square test.}

\item{trials}{The total number of trials. Required for Chi-Square test.}

\item{var_names}{a character vector of variables that can be combined to create
distinct names for each row and column.}
}
\value{
A square, symmetrical, with a length and width equal the number of rows in the data frame.
    Each cell in the matrix contains the results of the significance test from the row in the original
    dataframe represented by the column, and the row represented by the row in the matrix.
}
\description{
This function returns a square symetrical matrix of of all significance tests for all combinations of
values. It can calculate either a z-score or a p-value from a Chi-Square test of proportions.
The matrix length and width equal the number of rows in the data frame.
}
\details{
The z-score formula comes from: U.S. Census Bureau, A Compass for Understanding and Using ACS Survey Data, A-18
The Chi-Square test of proportions uses `prop.test`
}
\examples{
df <- data.frame(year = c(2016, 2017),
                 geo_description = c('Forsyth County, NC', 'Guilford County, NC'),
                 estimate = c(1,2),
                 se = c(.2, .3),
                 success = c(10, 12),
                 trials = c(15, 19))

# Z score test
ff_sigtest(data_frame = df, estimate = 'estimate', se = 'se',
           test = 'zscore', var_names = c('year', 'geo_description'))

# Chi-Square test
ff_sigtest(data_frame = df, estimate = 'estimate', success = 'success', trials = 'trials',
           test = 'chi-square', var_names = c('year', 'geo_description'))
}
