% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{ff_sigtest_kable}
\alias{ff_sigtest_kable}
\title{Pretty formatting of significance testing tables with Kable}
\usage{
ff_sigtest_kable(sigtest_matrix, test = "zscore", table_name = NULL)
}
\arguments{
\item{sigtest_matrix}{A matrix produced by `ff_sigtest`.}

\item{test}{The significance test to conduct. Either "zscore" or "chi-square". Defaults to 'zscore'.
Test must be the same test used in `ff_sigtest`.}

\item{table_name}{A string of characters representing the table name. It is displayed above the table.
Defautls to no table name.}
}
\value{
An html-based Kable table.
}
\description{
This function elegantly formats significance testing matrices produces
by `ff_sigtest`. It bolds cells that are statistically significant, only
keeps rows for Forsyth County, and makes other minor stylistic changes.
}
\examples{
df <- data.frame(year = c(2016, 2017),
                 geo_description = c('Forsyth County, NC', 'Guilford County, NC'),
                 estimate = c(1,2),
                 se = c(.2, .3),
                 success = c(10, 12),
                 trials = c(15, 19))

# Use ff_sigtest to create significance testing matrix
ff_sigtest(data_frame = df, estimate = 'estimate', se = 'se',
           test = 'zscore', var_names = c('year', 'geo_description')) \%>\%
           # Pipe significance testing matrix into function creating kable table
           ff_sigtest_kable(sigtest_matrix = ., test = 'zscore', table_name = 'Example table')
}
