% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{ff_spread_dual}
\alias{ff_spread_dual}
\title{Spread columns when there are two value columns that need to be retained.}
\usage{
ff_spread_dual(df, key, value_a, value_b, join_cols)
}
\arguments{
\item{df}{Dataframe of ACS data}

\item{key}{Key column in dataframe as string (the column with the information we want to spread on)}

\item{value_a}{First value column we want to keep as string}

\item{value_b}{Second value column we want to keep as string}

\item{join_cols}{Each value column is spread individually, and then both new spread datasets are joined.
`join_cols` signifies which column to join the to datasets on.}
}
\value{
A dataframe with seperate columns for each of the values spread on, and value_a and value_b.
The final dataframe will contain all the join columns and value columns totaling the number of items
to spread on times two (two different values that are kept when spreading).
}
\description{
The `spread` function only takes one value column.  But, we often need to spread a dataset
into wide-form, but with two value columns.  A example is when we want to spread a dataset,
but keep the estimate and standard error columns.  This function allows users to spread
a dataset and keep two value columns.
}
\examples{
ff_spread_dual(total, 'description', 'estimate', 'se', c('geo_description', 'year'))

}
