% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_census.R
\name{ff_table_names}
\alias{ff_table_names}
\title{Creates a vector of table names for all tables within a group.
For example, entering the parameter 'B20017' returns B20017, B20017A, B20017B, etc.}
\usage{
ff_table_names(table_number)
}
\description{
Creates a vector of table names for all tables within a group.
For example, entering the parameter 'B20017' returns B20017, B20017A, B20017B, etc.
}
