% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{ff_proportions}
\alias{ff_proportions}
\title{Difference and confidence interval of difference between two binomial observations}
\usage{
ff_proportions(successes, trials, rate_per_unit = 1)
}
\arguments{
\item{successes}{Vector of two numbers that represent the number of successes in each
of the observations.}

\item{trials}{Vector of two numbers that represent the number of trials in each
of the observation}

\item{rate_per_unit}{Integer used to calculate the rate per x number of people. For example, the crime rate
is the number of crimes per 100,000 people, so 100,000 would be entered. Defaults to 1, which is no adjustment.
Only used if format equals 'binomial'.}
}
\description{
This function calculates the percentile difference between two observations that
are binomial (true / false, yes / no, etc) in nature. It also calculates the confidence
interval of this difference. The function is used in `ff_estiamtes_ci`.
}
